site/monstermash/requirements.txt
Flask==0.9
Flask-Assets==0.8
Flask-Script==0.5.1
config==0.3.9
numpy==1.6.2
pyzmq==2.2.0.1
remix==2.0.1
requests==1.0.3
rethinkdb==1.5
